fix: cast total_exec_time to numeric for Postgres ROUND
Postgres has no ROUND(double precision, integer) overload, only
ROUND(numeric, integer) - discovered via Task 14's manual Postgres
smoke run, which failed with 'function round(double precision,
integer) does not exist' when fetching top statement digests.

+++ lib/bench/engines.rb
@@ -19,7 +19,7 @@ module Bench
     SQL
 
     POSTGRES_DIGEST_FETCH_SQL = <<~SQL.tr("\n", " ").freeze
-      SELECT query, calls, ROUND(total_exec_time, 1), rows
+      SELECT query, calls, ROUND(total_exec_time::numeric, 1), rows
       FROM pg_stat_statements
       WHERE dbid = (SELECT oid FROM pg_database WHERE datname = 'bench')
       ORDER BY total_exec_time DESC
